[FIX] 등록한 review 세션내 로그인한 사용자와 리뷰 등록 userID 동일해야 삭제 기능 추가

+++ src/main/java/ahchacha/ahchacha/service/ReviewService.java
@@ -70,62 +70,14 @@ public class ReviewService {
     }
 
 
-//    public Page<ReviewDto.ReviewResponseDto> getReviewsByItemIdHighScore(Long id, int page) {
-//        List<Sort.Order> sorts = new ArrayList<>();
-//        sorts.add(Sort.Order.desc("createdAt"));
-//        sorts.add(Sort.Order.desc("reviewScore"));
-//
-//        Pageable pageable = PageRequest.of(page - 1, 15, Sort.by(sorts));
-//        Page<Review> reviewPage = reviewRepository.findByItemId(id, pageable);
-//
-//        return ReviewDto.toDtoPage(reviewPage);
-//    }
-//    public Page<ReviewDto.ReviewResponseDto> getReviewsByItemIdLowScore(Long id, int page) {
-//        List<Sort.Order> sorts = new ArrayList<>();
-//        sorts.add(Sort.Order.asc("createdAt"));
-//        sorts.add(Sort.Order.desc("reviewScore"));
-//
-//        Pageable pageable = PageRequest.of(page - 1, 15, Sort.by(sorts));
-//        Page<Review> reviewPage = reviewRepository.findByItemId(id, pageable);
-//
-//        return ReviewDto.toDtoPage(reviewPage);
-//    }
-
-//    public Page<ReviewDto.ReviewResponseDto> getReviewsByUserId(Long id, int page) {
-//        List<Sort.Order> sorts = new ArrayList<>();
-//        sorts.add(Sort.Order.desc("createdAt"));
-//
-//        Pageable pageable = PageRequest.of(page - 1, 15, Sort.by(sorts));
-//        Page<Review> reviewPage = reviewRepository.findByUserId(id, pageable);
-//
-//        return ReviewDto.toDtoPage(reviewPage);
-//    }
-//
-//    public Page<ReviewDto.ReviewResponseDto> getReviewsByUserIdHighScore(Long id, int page) {
-//        List<Sort.Order> sorts = new ArrayList<>();
-//        sorts.add(Sort.Order.desc("createdAt"));
-//        sorts.add(Sort.Order.desc("reviewScore"));
-//
-//        Pageable pageable = PageRequest.of(page - 1, 15, Sort.by(sorts));
-//        Page<Review> reviewPage = reviewRepository.findByUserId(id, pageable);
-//
-//        return ReviewDto.toDtoPage(reviewPage);
-//    }
-//    public Page<ReviewDto.ReviewResponseDto> getReviewsByUserIdLowScore(Long id, int page) {
-//        List<Sort.Order> sorts = new ArrayList<>();
-//        sorts.add(Sort.Order.asc("createdAt"));
-//        sorts.add(Sort.Order.desc("reviewScore"));
-//
-//        Pageable pageable = PageRequest.of(page - 1, 15, Sort.by(sorts));
-//        Page<Review> reviewPage = reviewRepository.findByUserId(id, pageable);
-//
-//        return ReviewDto.toDtoPage(reviewPage);
-//    }
-
     @Transactional
-    public void deleteReview(Long reviewId) {
+    public void deleteReview(Long reviewId, User currentUser) {
         Review review = reviewRepository.findById(reviewId).orElseThrow(() ->
-                new IllegalArgumentException("Invalid review Id:"));
+                new IllegalArgumentException("Invalid review Id:" + reviewId));
+
+        if (!review.getUser().getId().equals(currentUser.getId())) {
+            throw new IllegalArgumentException("You do not have permission to delete this review.");
+        }
 
         reviewRepository.delete(review);
     }
